pvr: Some hacks to get things to build
The static_data.py change needs to be there since we're using the static
offset all the time.  These offsets are probably wrong (but we don't
expose any of them, so I don't think it matters)

The meson.build format changed a bit from the Chromium patches.

With these changes, I can use the EGL, gbm, libGLES libraries created
for kmscube to work.

Note that there is a difference in that it says it exposes
EGL_IMG_cl_image instead of EGL_IMG_context_priority - I don't know why,
but probably not critical.

Signed-off-by: Jonathan Bakker <xc-racer2@live.ca>